# docs/website/developer-guide/api/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Jetonomy REST API",
		"version": "1.9.0",
		"description": "Machine-readable OpenAPI 3.1 contract for the Jetonomy WordPress forum plugin (free core + Pro extensions). Built from the plugin's rest-api-reference.md, which was in turn generated by reading every register_rest_route() call in the source.\n\n**Authentication.** Two modes, both funnelled through REST_Auth::auth_mutation(): (1) Cookie + `X-WP-Nonce` header (the web app) - every mutation needs a valid `wp_rest` nonce; get a fresh one from `GET /auth/nonce`. (2) Application Passwords via HTTP Basic (mobile app + integrations), which skip the nonce check. Both modes enforce ban/silence and pending-email-verification gates.\n\n**Namespaces.** Almost everything is under `jetonomy/v1`. Two Pro surfaces live under `jetonomy-pro/v1` (marked with a per-operation `servers` override): the Attachments routes (`/attachments*`) and the Site Announcements admin routes (`/site-announcements*`). The public announcement read (`/announcements/active`) stays on `jetonomy/v1`.\n\n**Private-community mode.** When Settings -> Community Access is `private`, every `GET` requires login; a logged-out read returns `401 community_private`.\n\n**Pagination.** Offset-based (`limit`/`offset`) on most lists, cursor-based (`after`/`before`) on posts/replies/messages. Collection bodies are `{ data: [...], meta: { count, has_more, cursor_next, total, offset } }`.",
		"contact": {
			"name": "Wbcom Designs",
...
	"servers": [
...
	"security": [
		{
			"cookieAuth": [],
			"nonceAuth": []
		},
		{
			"appPassword": []
		}
	],
...
	"paths": {
...
		"/subscriptions": {
			"get": {
				"tags": [
					"Subscriptions"
				],
				"summary": "List my subscriptions",
				"operationId": "listSubscriptions",
				"description": "List the caller's subscriptions; each row includes the target's title/slug/exists (1.8.0). Login required.",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SubscriptionCollection"
								}
							}
						}
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					}
				},
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer"
						}
					},
					{
						"name": "offset",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer"
						}
					}
				]
			},
			"post": {
				"tags": [
					"Subscriptions"
				],
				"summary": "Subscribe",
				"operationId": "createSubscription",
				"description": "Subscribe to a space or post. Login required.",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Subscription"
								}
							}
						}
					},
					"400": {
						"$ref": "#/components/responses/BadRequest"
					},
					"401": {
						"$ref": "#/components/responses/Unauthorized"
					}
				},
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"object_type": {
										"type": "string",
										"enum": [
											"space",
											"post"
										]
									},
									"object_id": {
										"type": "integer"
									},
									"via": {
										"type": "string",
										"description": "Delivery channel"
									}
								},
								"required": [
									"object_type",
									"object_id"
								]
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "wordpress_logged_in",
				"description": "WordPress logged-in session cookie (the web app)."
			},
			"nonceAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-WP-Nonce",
				"description": "wp_rest nonce, required on every mutation for cookie-authenticated requests. Refresh via GET /auth/nonce."
			},
			"appPassword": {
				"type": "http",
				"scheme": "basic",
				"description": "WordPress Application Password sent as HTTP Basic auth (mobile app + external integrations). Skips the cookie-nonce check."
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "object",
						"properties": {
							"status": {
								"type": "integer"
							}
						}
					}
				},
				"required": [
					"code",
					"message"
				]
			},
			"PaginationMeta": {
				"type": "object",
				"properties": {
					"count": {
						"type": "integer"
					},
					"has_more": {
						"type": "boolean"
					},
					"cursor_next": {
						"type": [
							"string",
							"null"
						]
					},
					"total": {
						"type": "integer"
					},
					"offset": {
						"type": "integer"
					}
				}
			},
...
			"Subscription": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer"
					},
					"object_type": {
						"type": "string",
						"enum": [
							"space",
							"post"
						]
					},
					"object_id": {
						"type": "integer"
					},
					"via": {
						"type": [
							"string",
							"null"
						]
					},
					"title": {
						"type": [
							"string",
							"null"
						]
					},
					"slug": {
						"type": [
							"string",
							"null"
						]
					},
					"exists": {
						"type": "boolean"
					}
				}
			},
			"SubscriptionCollection": {
				"type": "object",
				"properties": {
					"data": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Subscription"
						}
					},
					"meta": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				}
			},
...
		},
		"responses": {
			"BadRequest": {
				"description": "Bad request / validation error (invalid or missing parameters).",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
			"Unauthorized": {
				"description": "Authentication required, or community is in private mode (401 community_private) for a logged-out read.",
				"content": {
					"application/json": {
						"schema": {
							"$ref": "#/components/schemas/Error"
						}
					}
				}
			},
...
		}
	}
}